See description!
Magic Data is no longer stored in Bank C, but is on Bank 4 (for now) and moved to $7000 ram for access.
Several placeholder spells have been added for skill, item, and enemy use.
A blinking _ has been added to the naming screen, and several tweaks there to help with re-doing a character's name.
The ! shown in shops next to characters who have the selected item equipped has been converted to sprites to speed up the shop menus--there was a delay between the highlight and the cursor moving to the next item, which is now only noticable to me at %50 emulation speed. I believe this should also stop the music from slowing down when cycling through items, but I can't confirm that at 5 AM.