Build.bat and nes.cfg re-write!
Parry changed to activate on turn start rather than selecting the option (no more selecting it, turning it on, then going back to choose another action!)
RAM initalized to 0 on reset.
$20 bytes saved in RAM that are no longer used to clear the invisible tiles below the battle windows.
build.bat and nes.cfg changed to better create .dbg files for Mesen! Thanks to Sour and the nesdev folks for helping sort that out.