MUSIC DRIVER UPDATE!
Cleaned up the Score with some Constants to hopefully make it easier to edit. Will also make it easier to adjust what each byte of score data does, so that a "Loop forever" command doesn't have to be "$D0" if sweeping changes need to happen.

Changed how the score is read and branches to different small routines to make it easier to navigate.

Changed the formate of Zero Page RAM usage for each instrument, so that the 16+7 bytes are now 14 bytes each, making room for possible future vibrato effects!